Start funcional
Si están disponibles se arranca el número de máquinas indicado, si no aparece un mensaje informando de que no se puede completar la acción

+++ pfinalp1.py
@@ -46,12 +46,18 @@ def start(numerodemaquinas):	#Configuracion de host
 	configuracionEscenario()
 	#Comprobacion de existencia de las maquinas
 	checkServers(numerodemaquinas)
-	if len(servidores_creados) - 2 < numerodemaquinas:
-		print("No existen las maquinas que desea arrancar, creelas mediante la opcion create num_servidores")
+	c1_lb_exist = False
+	servidores_exist = False
+	if servidores_creados.count("c1") and servidores_creados.count("lb"):
+		c1_lb_exist = True
+	if not len(servidores_creados)-2 < numerodemaquinas:
+		servidores_exist = True
+	if c1_lb_exist == True and servidores_exist == True:
+		arrancar(numerodemaquinas)
 	else:
-		#Comprobamos que maquinas estan arrancadas ya, para no arrancarlas 2 veces
+		print ("No se dan las condiciones correctas para arrancar las maquinas solicitadas")
 
-	return
+
 
 def stop():
 	return
@@ -134,12 +140,17 @@ def configuracionEscenario():
 	os.system("sudo ip route add 10.0.0.0/16 via 10.0.1.1")
 	os.system('HOME=/mnt/tmp sudo virt-manager')
 	return
-def checkArrancados(numeroaarrancar): #Rellena una lista con los servidores que ya han sido arrancados
+def arrancada(hostname):
+	if os.system("sudo virsh list | grep -q "+hostname) == 0:
+		return 1
+	else:
+		return 0
+'''def checkArrancados(numeroaarrancar): #Rellena una lista con los servidores que ya han sido arrancados
 	numero = numeroaarrancar
 	numeroserversarrancados = 0
 	for i in servidores_creados:
 		if os.system("sudo virsh list | grep -q "+i) == 0:
-			
+			return
 		else:
 			servidores_a_arrancar.append(i)
 	return
@@ -154,11 +165,118 @@ def servidoresAArrancar(numerodemaquinas): #Rellena una lista con los nombres de
 			else:
 				servidores_a_arrancar.append(i)
 				numeroservacrear = numeroservacrear - 1
-	return
-def arranca():#Arranca los servidores pendientes 
-	return
-
-			
+	return'''
+#Arranca los servidores pendientes 
+def arrancar(numerodemaquinas):
+	numeroarrancadas = numerodemaquinas
+	arranca("host")
+	if arrancada("c1") == 1:
+		print "Maquina virtual c1 esta arrancada"
+	else:
+		arranca("c1")
+	if arrancada("lb") == 1:
+		print "Maquina virtual lb esta arrancada"
+	else:
+		arranca("lb")
+	for i in servidores_creados:
+		if arrancada(i) == 1:
+			print "Maquina virtual "+i+" esta arrancada"
+			numeroarrancadas = numeroarrancadas -1
+		else:
+			arranca(i)
+			numeroarrancadas = numeroarrancadas -1
+def arranca(name):
+	if (name != "host"):
+		os.chdir("/mnt/tmp/pfinal")
+		os.system('sudo vnx_mount_rootfs -s -r '+name+'.qcow2 mnt')
+		os.chdir("/mnt/tmp/pfinal/mnt/etc")
+		dochostname = open("hostname", "w")
+		dochostname.write(name)
+		dochostname.close()
+		hosts(name)
+		os.chdir("/mnt/tmp/pfinal/mnt/etc/network")
+		interface(name)
+		os.chdir("/mnt/tmp/pfinal")
+		os.system('sudo vnx_mount_rootfs -u mnt')
+		if subprocess.call(["sudo","virsh","create",name+".xml"]) != 0:
+			os.chdir("/mnt/tmp/pfinal")
+			os.system('sudo virsh create '+name+'.xml')
+			terminal = "sudo virsh console "+name
+			p1 = Popen(['xterm', '-e', terminal])
+	else:
+		os.system("sudo ifconfig LAN1 10.0.1.3/24")
+		os.system("sudo ip route add 10.0.0.0/16 via 10.0.1.1")
+def hosts(name):
+	dochosts = open("hosts", "w")
+	dochosts.write("127.0.0.1	localhost"+"\n")
+	dochosts.close()
+	dochosts = open("hosts", "a")
+	dochosts.write("127.0.1.1	"+name+"\n")
+	dochosts.write("# The following lines are desirable for IPv6 capable hosts"+"\n")
+	dochosts.write("::1     localhost ip6-localhost ip6-loopback"+"\n")
+	dochosts.write("ff02::1 ip6-allnodes"+"\n")
+	dochosts.write("ff02::2 ip6-allrouters")
+	dochosts.close()
+def interface(name):
+	if  (name != "c1" and name != "lb"):
+		docinterface = open("interfaces", "w")
+		docinterface.write("# The loopback network interface"+"\n")
+		docinterface.close()
+		docinterface2 = open("interfaces", "a")
+		docinterface2.write("auto lo"+"\n")
+		docinterface2.write("iface lo inet loopback"+"\n")
+		docinterface2.write("# The primary network interface"+"\n")
+		docinterface2.write("auto eth0"+"\n")
+		docinterface2.write("iface eth0 inet static"+"\n")
+		docinterface2.write("address 10.0.2.1"+str(name[1])+"\n")
+		docinterface2.write("netmask 255.255.255.0"+"\n")
+		docinterface2.write("network 10.0.2.0"+"\n")
+		docinterface2.write("broadcast 10.0.2.255"+"\n")
+		docinterface2.write("gateway 10.0.2.1"+"\n")
+		docinterface2.close()
+	elif name == "c1":
+		docinterface = open("interfaces", "w")
+		docinterface.write("# The loopback network interface"+"\n")
+		docinterface.close()
+		docinterface2 = open("interfaces", "a")
+		docinterface2.write("auto lo"+"\n")
+		docinterface2.write("iface lo inet loopback"+"\n")
+		docinterface2.write("# The primary network interface"+"\n")
+		docinterface2.write("auto eth0"+"\n")
+		docinterface2.write("iface eth0 inet static"+"\n")
+		docinterface2.write("address 10.0.1.2"+"\n")
+		docinterface2.write("netmask 255.255.255.0"+"\n")
+		docinterface2.write("network 10.0.1.0"+"\n")
+		docinterface2.write("broadcast 10.0.1.255"+"\n")
+		docinterface2.write("gateway 10.0.1.1"+"\n")
+		docinterface2.close()
+	elif name == "lb":
+		docinterface = open("interfaces", "w")
+		docinterface.write("# The loopback network interface"+"\n")
+		docinterface.close()
+		docinterface2 = open("interfaces", "a")
+		docinterface2.write("auto lo"+"\n")
+		docinterface2.write("iface lo inet loopback"+"\n")
+		docinterface2.write("# The primary network interface"+"\n")
+		docinterface2.write("auto eth0"+"\n")
+		docinterface2.write("iface eth0 inet static"+"\n")
+		docinterface2.write("address 10.0.1.1"+"\n")
+		docinterface2.write("netmask 255.255.255.0"+"\n")
+		docinterface2.write("network 10.0.1.0"+"\n")
+		docinterface2.write("broadcast 10.0.1.255"+"\n")
+		docinterface2.write("gateway 10.0.1.1"+"\n")
+		docinterface2.write("# The secundary network interface"+"\n")
+		docinterface2.write("auto eth1"+"\n")
+		docinterface2.write("iface eth1 inet static"+"\n")
+		docinterface2.write("address 10.0.2.1"+"\n")
+		docinterface2.write("netmask 255.255.255.0"+"\n")
+		docinterface2.write("network 10.0.2.0"+"\n")
+		docinterface2.write("broadcast 10.0.2.255"+"\n")
+		docinterface2.write("gateway 10.0.2.1"+"\n")
+		docinterface2.close()
+		os.system("sed -i 's/#net.ipv4.ip_forward=1/net.ipv4.ip_forward=1/' /mnt/tmp/pfinal/mnt/etc/sysctl.conf")
+	else:
+		print "Solamente se pueden arrancar 7 maquinas virtuales como maximo"				
 #Comienzo de ejecucion
 
 #Lectura del comando:
@@ -183,10 +301,12 @@ elif len(parametros_comando) == 3: #Si se define la orden y los parametros corre
 			else:
 				os.mkdir("/mnt/tmp/pfinal")
 				os.chdir("/mnt/tmp/pfinal")
-			if not os.path.exists("cdps-vm-base-p3.qcow2.bz2"):
+			if not os.path.exists("cdps-vm-base-p3.qcow2.bz2") and not os.path.exists("cdps-vm-base-p3.qcow2"):
 				os.system("cp /mnt/vnx/repo/cdps/cdps-vm-base-p3.qcow2.bz2 .")
 				os.system("bunzip2 cdps-vm-base-p3.qcow2.bz2")
 			os.system("cp /mnt/vnx/repo/cdps/plantilla-vm-p3.xml .")
+			if not os.path.exists("/mnt/tmp/pfinal/mnt"):
+				os.mkdir("/mnt/tmp/pfinal/mnt")
 			print("Va bien")
 		else:
 			print("Parametro incorrecto")
